requirements.txt
streamlit
huggingface_hub
byaldi
transformers
Pillow
torch
openai
python-dotenv
